% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfl.R
\name{tfl_arrivals}
\alias{tfl_arrivals}
\title{Get live arrivals for a TfL line or stop}
\usage{
tfl_arrivals(line_id, stop_id = NULL)
}
\arguments{
\item{line_id}{Line ID (e.g. "victoria", "central", "northern")}

\item{stop_id}{Optional stop/station NAPTAN ID to filter arrivals}
}
\value{
tibble: id, line_name, station_name, destination, towards,
expected_arrival, time_to_station, platform_name, direction
}
\description{
Get live arrivals for a TfL line or stop
}
