% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfl.R
\name{tfl_lines}
\alias{tfl_lines}
\title{Get TfL lines by mode}
\usage{
tfl_lines(mode = "tube")
}
\arguments{
\item{mode}{Transport mode: "tube", "bus", "dlr", "overground",
"elizabeth-line", "tram", "cable-car" (default "tube")}
}
\value{
tibble: id, name, mode, status, reason
}
\description{
Get TfL lines by mode
}
