% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfl.R
\name{tfl_stops}
\alias{tfl_stops}
\title{Search TfL stops/stations}
\usage{
tfl_stops(query)
}
\arguments{
\item{query}{Search term (e.g. "Kings Cross", "Victoria")}
}
\value{
tibble: id, name, lat, lon, modes
}
\description{
Search TfL stops/stations
}
